refact: 폴더구조 변경 및 주석 추가
