+++ src/app.module.ts
@@ -4,6 +4,7 @@ import { typeORMConfig } from './configs/typeorm.config';
 import { UserModule } from './user/user.module';
 import { PostModule } from './post/post.module';
 import { StatisticsModule } from './statistics/statistics.module';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
@@ -11,6 +12,7 @@ import { StatisticsModule } from './statistics/statistics.module';
     UserModule,
     PostModule,
     StatisticsModule,
+    AuthModule,
   ],
 })
 export class AppModule {}
