+++ src/auth/auth.module.ts
@@ -1,0 +1,31 @@
+import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
+import * as config from 'config';
+import { EmailService } from './email.service';
+import { JwtStrategy } from './jwt.strategy';
+import { TypeOrmExModule } from 'src/common/decorator/typeorm-ex.module';
+import { AuthService } from './auth.service';
+import { AuthController } from './auth.controller';
+import { UserRepository } from 'src/user/user.repository';
+import { UserModule } from 'src/user/user.module';
+
+const jwtConfig = config.get('jwt');
+
+@Module({
+  imports: [
+    PassportModule.register({ defaultStrategy: 'jwt' }),
+    JwtModule.register({
+      secret: process.env.JWT_SECRET || jwtConfig.secret,
+      signOptions: {
+        expiresIn: jwtConfig.expiresIn,
+      },
+    }),
+    UserModule,
+    TypeOrmExModule.forCustomRepository([UserRepository]),
+  ],
+  controllers: [AuthController],
+  providers: [AuthService, JwtStrategy, EmailService],
+  exports: [JwtStrategy, PassportModule],
+})
+export class AuthModule {}
