+++ src/user/user.module.ts
@@ -1,15 +1,12 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRepository } from './user.repository';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { TypeOrmExModule } from '../common/decorator/typeorm-ex.module';
+import { JwtStrategy } from '../auth/jwt.strategy';
 import { PassportModule } from '@nestjs/passport';
-import * as config from 'config';
 import { JwtModule } from '@nestjs/jwt';
-import { JwtStrategy } from './jwt.strategy';
-import { User } from './entities/user.entity';
-import { EmailService } from './email.service';
+import * as config from 'config';
 
 const jwtConfig = config.get('jwt');
 
@@ -22,11 +19,10 @@ const jwtConfig = config.get('jwt');
         expiresIn: jwtConfig.expiresIn,
       },
     }),
-    TypeOrmModule.forFeature([User]),
     TypeOrmExModule.forCustomRepository([UserRepository]),
   ],
   controllers: [UserController],
-  providers: [UserService, JwtStrategy, EmailService],
-  exports: [JwtStrategy, PassportModule],
+  providers: [UserService, JwtStrategy],
+  exports: [TypeOrmExModule],
 })
 export class UserModule {}
